requirements.txt
# SRE AI Agent — Dependencies
# ─────────────────────────────

# LangGraph & LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-aws>=0.2.0

# AWS
boto3>=1.35.0

# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Datadog Observability
ddtrace>=2.14.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.0.0

# Configuration & Utilities
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
